Remove Payload class
