+++ architect/systems/__init__.py
@@ -1,2 +1,2 @@
 from .system import System  # isort: skip
-from . import payloads, satellites
+from . import satellites
